+++ chocolatey/tools/chocolateyInstall.ps1
@@ -1,4 +1,4 @@
 ﻿$ErrorActionPreference = 'Stop'
 
 $toolsPath = Split-Path $MyInvocation.MyCommand.Definition
-& "$toolsPath/install.ps1" -module_path $toolsPath/AU | Out-Null
+& "$toolsPath/install.ps1"